requirements.txt
albumentations==2.0.7
asposestorage==1.0.2
feature_engine==1.8.3
ignite==1.1.0
imgaug==0.4.0
lightning==2.5.1.post0
matplotlib==3.10.3
monai==1.4.0
mpire==2.10.2
nibabel==5.3.2
nltk==3.9.1
numpy==2.2.6
omegaconf==2.3.0
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.2.1
pyradiomics==3.0.1
scikit_learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
shap==0.47.2
SimpleITK==2.4.1
SimpleITK==2.5.0
skimage==0.0
timm==1.0.15
torch==2.6.0
torchmetrics==1.7.1
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
